Move `qml.cond` to `ops.op_math` (#4860)
* Moved `condition.py` from the `transforms` folder to `ops/op_math`.
`qml.transforms.Conditional` is now available as `qml.ops.Conditional`.
* Updated docs accordingly.
