+++ tests/ops/op_math/test_condition.py
@@ -26,7 +26,7 @@ files.
 import pytest
 
 import pennylane as qml
-from pennylane.transforms.condition import ConditionalTransformError
+from pennylane.ops.op_math.condition import ConditionalTransformError
 
 terminal_meas = [
     qml.probs(wires=[1, 0]),
@@ -64,16 +64,16 @@ class TestCond:
         assert len(ops) == 4
         assert ops[0].return_type == qml.measurements.MidMeasure
 
-        assert isinstance(ops[1], qml.transforms.condition.Conditional)
+        assert isinstance(ops[1], qml.ops.Conditional)
         assert isinstance(ops[1].then_op, qml.PauliX)
         assert ops[1].then_op.wires == target_wire
 
-        assert isinstance(ops[2], qml.transforms.condition.Conditional)
+        assert isinstance(ops[2], qml.ops.Conditional)
         assert isinstance(ops[2].then_op, qml.RY)
         assert ops[2].then_op.wires == target_wire
         assert ops[2].then_op.data == (r,)
 
-        assert isinstance(ops[3], qml.transforms.condition.Conditional)
+        assert isinstance(ops[3], qml.ops.Conditional)
         assert isinstance(ops[3].then_op, qml.PauliZ)
         assert ops[3].then_op.wires == target_wire
 
@@ -129,20 +129,20 @@ class TestCond:
 
         assert ops[0].return_type == qml.measurements.MidMeasure
 
-        assert isinstance(ops[1], qml.transforms.condition.Conditional)
+        assert isinstance(ops[1], qml.ops.Conditional)
         assert isinstance(ops[1].then_op, qml.PauliX)
         assert ops[1].then_op.wires == target_wire
 
-        assert isinstance(ops[2], qml.transforms.condition.Conditional)
+        assert isinstance(ops[2], qml.ops.Conditional)
         assert isinstance(ops[2].then_op, qml.RY)
         assert ops[2].then_op.wires == target_wire
         assert ops[2].then_op.data == (r,)
 
-        assert isinstance(ops[3], qml.transforms.condition.Conditional)
+        assert isinstance(ops[3], qml.ops.Conditional)
         assert isinstance(ops[3].then_op, qml.PauliZ)
         assert ops[3].then_op.wires == target_wire
 
-        assert isinstance(ops[4], qml.transforms.condition.Conditional)
+        assert isinstance(ops[4], qml.ops.Conditional)
         assert isinstance(ops[4].then_op, qml.PauliY)
         assert ops[4].then_op.wires == target_wire
 
@@ -224,12 +224,12 @@ class TestOtherTransforms:
         assert len(ops) == 3
         assert ops[0].return_type == qml.measurements.MidMeasure
 
-        assert isinstance(ops[1], qml.transforms.condition.Conditional)
+        assert isinstance(ops[1], qml.ops.Conditional)
         assert isinstance(ops[1].then_op, qml.ops.op_math.Adjoint)
         assert isinstance(ops[1].then_op.base, qml.RX)
         assert ops[1].then_op.wires == target_wire
 
-        assert isinstance(ops[2], qml.transforms.condition.Conditional)
+        assert isinstance(ops[2], qml.ops.Conditional)
         assert isinstance(ops[2].then_op, qml.RX)
         assert ops[2].then_op.data == (r,)
         assert ops[2].then_op.wires == target_wire
@@ -253,11 +253,11 @@ class TestOtherTransforms:
         assert len(ops) == 3
         assert ops[0].return_type == qml.measurements.MidMeasure
 
-        assert isinstance(ops[1], qml.transforms.condition.Conditional)
+        assert isinstance(ops[1], qml.ops.Conditional)
         assert isinstance(ops[1].then_op, qml.ops.op_math.Controlled)
         assert qml.equal(ops[1].then_op.base, qml.RX(r, wires=2))
 
-        assert isinstance(ops[2], qml.transforms.condition.Conditional)
+        assert isinstance(ops[2], qml.ops.Conditional)
         assert isinstance(ops[2].then_op, qml.ops.op_math.Controlled)
         assert qml.equal(ops[2].then_op.base, qml.RY(r, wires=2))
 
@@ -281,11 +281,11 @@ class TestOtherTransforms:
         assert ops[0].return_type == qml.measurements.MidMeasure
 
         assert isinstance(ops[1], qml.ops.op_math.Controlled)
-        assert isinstance(ops[1].base, qml.transforms.condition.Conditional)
+        assert isinstance(ops[1].base, qml.ops.Conditional)
         assert qml.equal(ops[1].base.then_op, qml.RX(1.234, wires=0))
 
         assert isinstance(ops[2], qml.ops.op_math.Controlled)
-        assert isinstance(ops[2].base, qml.transforms.condition.Conditional)
+        assert isinstance(ops[2].base, qml.ops.Conditional)
         assert qml.equal(ops[2].base.then_op, qml.RY(r, wires=0))
 
         assert len(tape.measurements) == 1
